1.新增个性化消息输出框 2.指令解析到算法完成

+++ GAnalysis/Command/mycommandrules.h
@@ -80,10 +80,10 @@ public:
     gPoint();
     ~gPoint();
 public:
-    double x;
-    double y;
-    double i;
-    double j;
+    int x;
+    int y;
+    int i;
+    int j;
 };
 
 //指令结构体：储存指令代表的信息
